projects/telegram5/requirements.txt
ConfigParser
Cython
Flask
HTMLParser
IPython
api
argparse
ast
asyncio
attr
auth
calendar
config
configparser
cryptography
csv
ctypes
database
dataclasses
datetime
dis
distutils
dl
docutils
email
enum
flask_login
flask_restful
flask_sqlalchemy
functools
google
hashlib
hmac
html
http
importlib
importlib_metadata
io
ipaddress
ipywidgets
jinja2
jnius
keyring
lockfile
logging
media
messaging
numbers
numpy
packaging
pathlib
pillow
pip
pyparsing
railroad
routes
security
select
setuptools
socks
sphinx
ssl
thread
time
toml
tornado
trove_classifiers
truststore
typing
ui
unittest2
uuid
wheel
xmlrpclib
zipp
